Updated the RawData and Evidence to use LONG time format instead of INT, adaptions to some agnents

+++ netrank/src/main/java/netrank/FileSource.java
@@ -58,7 +58,7 @@ public class FileSource implements DataAgentInterface {
         }
 
         int start_time = 0;
-        int first_data_time = 0;
+        long first_data_time = 0;
 
         pattern = Pattern.compile(regex);
